Fix: fixed issues due to refactor
